#77 refact : GroupeNodeList 기능을 NodeList에서 사용 가능하도록
