+++ app/_components/Network/styles/NodeList.module.css
@@ -1,29 +1,54 @@
 /* NodeList.module.css */
 .nodeList {
-    list-style-type: none;
-    padding: 0;
-    width: 100%;
-  }
-  
-  .nodeItem {
-    cursor: pointer;
-    padding: 8px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    margin-bottom: 5px;
-    transition: background-color 0.3s, color 0.3s;
-    width: 100%;
-    box-sizing: border-box;
-  }
-  
-  .nodeItemSelected {
-    background-color: #96a0fe;
-    color: #fff;
-    border: 1px solid #96a0fe;
-  }
-  
-  .nodeItemNotSelected {
-    background-color: #fff;
-    color: #333;
-  }
-  
+  list-style-type: none;
+  padding: 0;
+  width: 100%;
+}
+
+.nodeItem {
+  cursor: pointer;
+  padding: 8px;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  margin-bottom: 5px;
+  transition: background-color 0.3s, color 0.3s;
+  width: 100%;
+  box-sizing: border-box;
+}
+
+.nodeItemSelected {
+  background-color: #96a0fe;
+  color: #fff;
+  border: 1px solid #96a0fe;
+}
+
+.nodeItemNotSelected {
+  background-color: #fff;
+  color: #333;
+}
+
+.groupHeader {
+  cursor: pointer;
+  padding: 10px;
+  background-color: #fff;
+  color: #96a0fe;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  margin-bottom: 5px;
+  transition: background-color 0.3s, color 0.3s;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.groupHeaderExpanded {
+  background-color: #96a0fe;
+  color: #fff;
+}
+
+.nodeList {
+  list-style-type: none;
+  padding: 0 10px;
+  margin: 0;
+  transition: max-height 0.3s;
+}
